Create Business Review
Script to generate business commercial KPIs and track them against the targets set(both monthly and weekly targets). The kpis are tracked in the current week( starting Monday) , the previous week( starting Monday) and also in the Month to Date period. The main purpose of the script is to help guide the business on it's progress and help push initiatives on which areas are to be improved to meet set targets.